@@ -87,10 +87,11 @@
     "toastify-js": "^1.9.0",
     "toastr": "^2.1.4",
     "url-loader": "^1.1.2",
-    "vue": "2.5.17",
-    "vue-loader": "^15.3.0",
+    "vue": "2.7.16",
+    "vue-loader": "^15.10.0",
+    "vue-router": "3.6.5",
     "vue-style-loader": "^4.1.2",
-    "vue-template-compiler": "2.5.17",
+    "vue-template-compiler": "2.7.16",
     "vue2-highcharts": "^1.2.5",
     "vuedraggable": "^2.16.0",
     "vueify": "^9.4.1",
